refactor[api-server]: 외부 의존성을 줄인 Embedded RedisConfig 적용
